Store DexFile instances in Architecture["Smali"]
From Architecture plugins (after `__init__` has been called),
`Architecture["Smali"]` and `self` are two different instances, and
BinaryView plugins have access to the `Architecture["Smali"]` one. This
eliminates the need for writing data to the disk.

Note that this instance is still shared between all tabs, so it does not
affect the inability to load two different Dex files in one window.

The code for writing JSON files for debugging purposes is left in but
commented out. It would be nice to hook that up to a GUI button or a
function from the console or something.

Close #8