Add initial support for multiple binja tabs
This uses UI context notifications to tell which file is in view and
load the relevant disassembly. This is only a heuristic, however,
because if you switch between files while things are still processing or
open two large files quickly, it's possible that it will get confused.
This is a lot better than it was before, but it's not foolproof. I don't
think it's possible to get to 100% just using UI context notifications,
but the heuristics could still at least be improved.

During testing, things were pretty ok, so it's good enough for now.

Close #14 (will open new issue for improvement)